+++ examples/resources/elestio_discourse/resource.tf
@@ -1,6 +1,6 @@
 resource "elestio_discourse" "example" {
   project_id    = "2500"
-  version       = "3.4.7-debian-12-r0"
+  version       = "release-stable"
   provider_name = "netcup"
   datacenter    = "nbg"
   server_type   = "MEDIUM-2C-4G"