+++ examples/resources/elestio_getateam/import.sh
@@ -1,0 +1,2 @@
+# Import a GetATeam service by specifying the Project ID it belongs to, and the GetATeam service ID (spaced by a comma).
+terraform import elestio_getateam.my_getateam project_id,getateam_service_id