+++ examples/resources/elestio_pgvector/import.sh
@@ -1,0 +1,2 @@
+# Import a pgvector service by specifying the Project ID it belongs to, and the pgvector service ID (spaced by a comma).
+terraform import elestio_pgvector.my_pgvector project_id,pgvector_service_id