+++ examples/resources/elestio_supabase/resource.tf
@@ -1,6 +1,6 @@
 resource "elestio_supabase" "example" {
   project_id    = "2500"
-  version       = "20241202-71e5240"
+  version       = "2025.11.24-sha-d990ae8"
   provider_name = "netcup"
   datacenter    = "nbg"
   server_type   = "MEDIUM-2C-4G"