+++ examples/resources/elestio_cryptomator/import.sh
@@ -1,0 +1,2 @@
+# Import a Cryptomator service by specifying the Project ID it belongs to, and the Cryptomator service ID (spaced by a comma).
+terraform import elestio_cryptomator.my_cryptomator project_id,cryptomator_service_id