+++ examples/resources/elestio_postgis/import.sh
@@ -1,0 +1,2 @@
+# Import a postGis service by specifying the Project ID it belongs to, and the postGis service ID (spaced by a comma).
+terraform import elestio_postgis.my_postgis project_id,postgis_service_id